requirements.txt
albumentations==0.4.3
python-dateutil 
pytz 
tqdm 
cmake 
filelock
diffusers
opencv-python==4.1.2.30
pudb==2019.2
invisible-watermark
imageio==2.9.0
imageio-ffmpeg==0.4.2
pytorch-lightning==1.6.1
omegaconf
test-tube>=0.7.5
streamlit>=0.73.1
einops==0.3.0
torch-fidelity==0.3.0
transformers==4.19.2
torchmetrics==0.6.0
kornia==0.6
simple-term-menu
gdown
av
wandb
lmdb
ipdb
tqdm
clean-fid
gradio
lpips
joblib
-e git+https://github.com/CompVis/taming-transformers.git@master#egg=taming-transformers
git+https://github.com/openai/CLIP # edited spec to match k-diffusion
git+https://github.com/facebookresearch/detectron2.git
git+https://github.com/crowsonkb/k-diffusion.git
-e .
